+++ propan/fastapi/nats/__init__.py
@@ -1,0 +1,3 @@
+from propan.fastapi.nats.router import NatsRouter
+
+__all__ = ("NatsRouter",)