+++ propan/fastapi/nats/router.py
@@ -1,0 +1,7 @@
+from propan import NatsBroker
+from propan.fastapi.core.router import PropanRouter
+
+
+class NatsRouter(PropanRouter):
+    broker_class = NatsBroker
+    broker: NatsBroker