+++ tests/brokers/nats/test_consume.py
@@ -1,0 +1,8 @@
+import pytest
+
+from tests.brokers.base.consume import BrokerConsumeTestcase
+
+
+@pytest.mark.nats
+class TestNatsConsume(BrokerConsumeTestcase):
+    pass