+++ tests/brokers/nats/test_rpc.py
@@ -1,0 +1,8 @@
+import pytest
+
+from tests.brokers.base.rpc import BrokerRPCTestcase
+
+
+@pytest.mark.nats
+class TestNatsRPC(BrokerRPCTestcase):
+    pass